Fix #66 by introducing mkExtraKindBinders  (#68)
* Refactor normalizeDecFor

Currently, normalizeDecFor is a chaotic mess of code duplication.
This attempts to consolidate as much code as possible into a
central function, `normalize'`. This paves the way for the main
event, which is the fix for #66.

* Fix #66 by introducing mkExtraKindBinders

normalizeDec might be given a declaration like this:

  data Foo :: Type -> Type

The TH AST will claim that this has no TyVarBndrs, but there actually
is one: it's hidden in the return kind! I define a mkExtraKindBinders
function to flush it out so that the DatatypeInfo reports it
properly.