Make freeVariablesWellScoped sort independently of unique order (#71)
Previously, `freeVariablesWellScoped` was performing a topological
sort on a graph constructed from the `graphFromEdges` function in the
`containers` library, which relies on the `Ord Name` instance to
work. (See #70 for why this is dodgy.) In this patch, I replace this
component of `freeVariablesWellScoped` with `scopedSort`, a
deterministic function which mimics the topological sorting without
relying on the `Ord Name` instance. (`scopedSort`'s implementation is
cargo-culted from the function of the same name in the GHC source
code.)

Fixes #70.