Support ForallVisT, fix #79 (#80)
`template-haskell-2.16.0.0` (bundled with GHC 8.10.1) introduces a
`ForallVisT` constructor to represent visible dependent
quantification. This patch updates the machinery in `th-abstraction`
to handle `ForallVisT`. Most of the lines of code added are
copy-pasted from the code that handles `ForallT`.

While writing this patch I noticed #79, so this patch fixes #79 as
well. glguy also noticed yet another bug in
https://github.com/glguy/th-abstraction/pull/80/files#r375448846,
so I have added a regression test for it as `t80Test`.