Explicitly mark modules as Safe or Trustworthy
I made a minor version bump, as here we actually have a small API change.
For older GHCs, modules are not Unsafe, but Trustworthy now.
So if someone want to be precise and write CPP-conditionals,
it is possible.

For the reference tagged-0.8.6.1 is on the edge, there visibly little
changed, though there Data.Proxy.TH module went through similar
transition.

(PVP doesn't mention Safe Haskell, we have to improvise).