repairVarKindsWith: Include binders from explicit return kinds
Due to an oversight, the implementation of `repairVarKindsWith` could
inadvertently truncate the value it returns when its `[TyVarBndr]` argument is
less than its `[Type]` argument, which leads to the confusing results seen
in #95. This can actually happen in practice when dealing with data family
declarations like `data family D :: Type -> Type`, where some of the kinds are
declared in the return kind rather than in the type variable binders.

Fortunately, we already have a `mkExtraKindBinders` function just for this very
purpose, so fixing this bug is a simple matter of using `mkExtraKindBinders` in
`repairVarKindsWith` to ensure that the `TyVarBndr` list is of equal length to
the `Type` list.

Fixes #95.