Compute datatypeInstVars correctly when return kind contains `forall`s
Previously, we were simply calling `freeVariablesWellScoped` on the explicit
return kind in a data type, but this gave incorrect results for data types like
this one:

```hs
data T :: forall k. k -> Type where ...
```

Here, we want `datatypeInstVars` to be `[k, (a :: k)]` (where `a` is fresh),
but we were instead getting `[(a :: k)]`. The problem was that `k` doesn't
occur free in `forall k. k -> Type`, so `th-abstraction` never realized that
`k` should be a part of `datatypeInstVars`.

We now compute `datatypeInstVars` in a more intelligent way by splitting apart
the explicit return kind using the new `mkExtraFunArgForalls` function. We then
pass the resulting `FunArgs` to `funArgTys` and compute the free variables of
the resulting `Type`s. By this point, all `forall`s have been split apart,
which ensures that we detect things like the `k` in `forall k. ...`.

Fixes #110.